Abstraction of user DTO and ResponseUserDTO. Added role to User.
